- Προσθήκη ανάθεσης σχολικής βιβλιοθήκης σε εκπ/κό σε επίπεδο σχολικής μονάδας, από τη σελίδα επεξεργασίας (issue #15) - Διόρθωση σφάλματος pagination σελίδας μονίμων

+++ dipedb.sql
@@ -747,7 +747,7 @@ CREATE TABLE IF NOT EXISTS `school` (
   `anenergo` tinyint(4) NOT NULL,
   `perif` int(11) NOT NULL,
   `systeg` int(11) NOT NULL,
-  `vivliothiki` tinyint(1) NOT NULL,
+  `vivliothiki` int(11) NOT NULL,
   `archive` text NOT NULL
 ) ENGINE=InnoDB AUTO_INCREMENT=405 DEFAULT CHARSET=greek;
